# docs/swagger.yaml
basePath: /
definitions:
  api.Response:
    properties:
      error:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
  catalog.Product:
    properties:
      category:
        type: string
      code:
        type: string
      price:
        type: number
      variants:
        items:
          $ref: '#/definitions/catalog.Variant'
        type: array
    type: object
  catalog.Variant:
    properties:
      name:
        type: string
      price:
        type: number
    type: object
  category.Category:
    properties:
      code:
        type: string
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: The API used to interact with the catalog, including products, variants
    and categories.
  title: Mytheresa
  version: "1.0"
paths:
  /catalog:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/catalog.Product'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      summary: Returns products in a paged format
      tags:
      - catalog
  /catalog/{code}:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/catalog.Product'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      summary: Returns a specific product from the catalog or none if not found
      tags:
      - catalog
  /categories:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/category.Category'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      summary: Get all available categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.Response'
      summary: Create a new category
      tags:
      - categories
swagger: "2.0"